common/tasks: split housekeeper and queue management tasks into separate files
